chatbot-app/backend/requirements.txt
Flask
flask-cors
requests
python-dotenv

